Exemples/requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
requests>=2.31.0

